+++ backend/packages/spend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/spend
 
+## 0.13.12
+
+### Patch Changes
+
+- Updated dependencies [413095f]
+  - @cat-factory/contracts@0.222.0
+  - @cat-factory/kernel@0.224.0
+
 ## 0.13.11
 
 ### Patch Changes